Refactorizando funciones de importación
